@@ -50,41 +42,42 @@
   "module": "index.js",
   "typings": "index.d.ts",
   "dependencies": {
-    "@angular/animations": "^4.4.7",
-    "@angular/common": "^4.4.7",
-    "@angular/compiler": "^4.4.7",
-    "@angular/compiler-cli": "^4.4.7",
-    "@angular/core": "^4.4.7",
-    "@angular/forms": "^4.4.7",
-    "@angular/platform-browser": "^4.4.7",
-    "@angular/platform-browser-dynamic": "^4.4.7",
-    "@angular/platform-server": "^4.4.7",
-    "@angular/tsc-wrapped": "^4.4.7"
+    "@angular/common": "~7.0.0",
+    "@angular/compiler": "~7.0.0",
+    "@angular/core": "~7.0.0",
+    "@angular/forms": "~7.0.0",
+    "@angular/platform-browser": "~7.0.0",
+    "@angular/platform-browser-dynamic": "~7.0.0",
+    "core-js": "^2.5.4",
+    "rxjs": "~6.3.3",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@types/jasmine": "^2.8.9",
-    "@types/karma": "^3.0.0",
-    "@types/karma-coverage": "^0.5.34",
-    "@types/karma-jasmine": "^0.0.30",
-    "@types/node": "^10.12.0",
+    "webpack": "^4.21.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "~7.0.1",
+    "@angular/compiler-cli": "~7.0.0",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
     "angular2-template-loader": "^0.6.2",
+    "codelyzer": "~4.5.0",
     "copyfiles": "^2.1.0",
-    "core-js": "^2.5.7",
-    "jasmine-core": "^3.2.1",
-    "karma": "^3.0.0",
-    "karma-jasmine": "^1.1.2",
-    "karma-phantomjs-launcher": "^1.0.4",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^3.0.5",
-    "phantomjs-prebuilt": "^2.1.16",
+    "protractor": "~5.4.0",
     "raw-loader": "^0.5.1",
-    "reflect-metadata": "^0.1.12",
-    "rxjs": "^5.5.11",
     "ts-loader": "^5.2.2",
-    "tslint": "^5.11.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
     "typedoc": "^0.13.0",
-    "typescript": "^2.9.2",
-    "webpack": "^4.21.0",
-    "zone.js": "^0.8.26"
+    "typescript": "~3.1.1"
   }
 }
