@@ -30,7 +30,13 @@
       colors: true,
       logLevel: config.LOG_INFO,
       autoWatch: false,
-      browsers: ['PhantomJS'],
+      browsers: ['Chrome'],
+      customLaunchers: {
+        ChromeHeadlessCI: {
+          base: 'ChromeHeadless',
+          flags: ['--no-sandbox']
+        }
+      },
       singleRun: true
     };
 
